commit do fim do dia 2 (stage falhou no anterior, vide descrição do mesmo)

+++ src/main/java/com/duayres/repository/IAgendamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.duayres.model.Agendamento;
 
-public interface IAgendamentoRepository extends JpaRepository<Agendamento, Integer> {
+public interface IAgendamentoRepository extends JpaRepository<Agendamento, Long> {
 
 }
