+++ src/main/java/com/duayres/repository/ILocalizacaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.duayres.model.Localizacao;
 
-public interface ILocalizacaoRepository extends JpaRepository<Localizacao, Integer> {
+public interface ILocalizacaoRepository extends JpaRepository<Localizacao, Long> {
 
 }
