+++ src/main/java/com/duayres/repository/ITipoDeEquipamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.duayres.model.TipoDeEquipamento;
 
-public interface ITipoDeEquipamentoRepository extends JpaRepository<TipoDeEquipamento, Integer> {
+public interface ITipoDeEquipamentoRepository extends JpaRepository<TipoDeEquipamento, Long> {
 
 }
